rename get_entities_popularity -> get_entities_popularity_cache
