+++ server/controllers/entities/popularity.js
@@ -2,7 +2,7 @@ const __ = require('config').universalPath
 const error_ = __.require('lib', 'error/error')
 const responses_ = __.require('lib', 'responses')
 const sanitize = __.require('lib', 'sanitize/sanitize')
-const getEntitiesPopularity = require('./lib/get_entities_popularity')
+const getEntitiesPopularityCache = require('./lib/get_entities_popularity_cache')
 
 const sanitization = {
   uris: {},
@@ -10,13 +10,13 @@ const sanitization = {
 }
 
 // TODO: when passing a refresh flag, return the old popularity value
-// instead of the default value, as getEntitiesPopularity would do
+// instead of the default value, as getEntitiesPopularityCache would do
 // for Wikidata entities
 module.exports = (req, res) => {
   sanitize(req, res, sanitization)
   .then(params => {
     const { uris, refresh } = params
-    return getEntitiesPopularity(uris, refresh)
+    return getEntitiesPopularityCache(uris, refresh)
   })
   .then(responses_.Wrap(res, 'scores'))
   .catch(error_.Handler(req, res))
